Bug: #68 Requirement: R2153, R2161
- Window/levelling of colour images no longer causes a crash
- Changed the ImageGraphic object hierarchy.  Now, we distinguish between grayscale and colour ImageGraphic objects for better separation of responsibilities.  See the ImageGraphics.cd class diagram for revised hierarchy.  This refactoring allows for improved code reuse when people want to create their own ImageGraphics and PresentationImages.
- The LUT pipeline was refactored to include the DICOM Presentation LUT Transformation and allow for the Colour Palette transformation in the future.  The output of the Presentation LUT (which MUST be the last LUT in the pipeline) is now an "int" instead of a "byte", where that "int" is actually an ARGB representation.  This allows support for pseudo colour maps for grayscale images.
- The wrapper to allow easier access to pixel data (formerly called PixelDataWrapper; now just PixelData) has now been incorporated into ImageGraphic.

+++ ImageViewer/Rendering/ImageInterpolator.cs
@@ -33,7 +33,7 @@ namespace ClearCanvas.ImageViewer.Rendering
 			int dstWidth,
 			int dstBytesPerPixel,
 			bool swapXY,
-			byte* pLutData,
+			int* pLutData,
 			bool isRGB,
 			bool isPlanar,
 			bool IsSigned);
